Добавляет контроллер для рендера событий
